Extract types into own package.
